Take the code of filesystem preparation out of container
For reusability
